revert: OAuth instance cache issue (#1492) (#1505)
This reverts commit 7506aa28f48e4fda58cfe50c360755b53c47db84.
