+++ commands/serve/web/hooks/index.js
@@ -3,4 +3,3 @@ export { useConnection } from './useConnection';
 export { useAppList } from './useAppList';
 export { useOpenApp } from './useOpenApp';
 export { useCachedConnections } from './useCachedConnections';
-export { useDeauthorizePrevOAuthInstance } from './useDeauthorizePrevOAuthInstance';
